Delete old NFT
