+++ .env
@@ -9,6 +9,6 @@ REACT_APP_INFURA_RINKEBY=rinkeby.infura.io/v3/35be3fddff544443acd2f8faa9134641
 REACT_APP_INFURA_GORLI=goerli.infura.io/v3/35be3fddff544443acd2f8faa9134641
 REACT_APP_LOGGLY_TOKEN=0c02fa85-a311-4c99-9b0b-102b79ef16c2
 
-E2E_METAMASK_SEED=
-E2E_METAMASK_ADDRESS=
-E2E_METAMASK_PASSWORD=
+E2E_METAMASK_SEED=away naive later first typical expect laundry kiwi similar jelly flower walnut
+E2E_METAMASK_ADDRESS=0x6c0A456c55E6485e7EF3540b0d6ceE6f54e21107
+E2E_METAMASK_PASSWORD=123bc6e035!#4
