+++ src/protocol/ethereum/featureReducer.tsx
@@ -5,7 +5,7 @@ import { useWeb3React } from '@web3-react/core'
 import { Reducer as NetworkReducer } from './network/Reducer'
 import { Reducer as UserReducer } from './user/Reducer'
 import { Reducer as ThreeBoxReducer } from './threeBox/Reducer'
-import { Reducer as NftReducer } from './nft-/Reducer'
+import { Reducer as NftReducer } from './nft/Reducer'
 import { Reducer as CustomContractReducer } from './customContract/Reducer'
 
 // Types
