--- src/protocol/ethereum/nft/single/NftSingleCustomImgField.tsx
@@ -1,6 +1,0 @@
-import React from 'react'
-
-export const NftSingleCustomField = ({ element, fieldData }) => {
-  element.innerHTML = fieldData
-  return null
-}
